+++ packages/forma-36-react-components/src/components/EntityList/EntityList/EntityList.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import cn from 'classnames';
 
-const styles = require('./EntityList.css');
+import styles from './EntityList.css';
 
 export type EntityListProps = {
   /**